+++ android/src/main/java/com/mediadevkit/fvp/FvpPlugin.java
@@ -1,11 +1,14 @@
 /*
- * Copyright (c) 2023 WangBin <wbsecg1 at gmail.com>
+ * Copyright (c) 2023-2025 WangBin <wbsecg1 at gmail.com>
  */
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 package com.mediadevkit.fvp;
 
+import android.content.Context;
+import android.content.pm.PackageManager;
 import android.graphics.SurfaceTexture;
+import android.os.Bundle;
 import android.util.Log;
 import android.view.Surface;
 
@@ -20,7 +23,9 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.view.TextureRegistry;
+import io.flutter.view.TextureRegistry.TextureEntry;
 import io.flutter.view.TextureRegistry.SurfaceTextureEntry;
+import io.flutter.view.TextureRegistry.SurfaceProducer;
 
 /** FvpPlugin */
 public class FvpPlugin implements FlutterPlugin, MethodCallHandler {
@@ -31,10 +36,23 @@ public class FvpPlugin implements FlutterPlugin, MethodCallHandler {
   private MethodChannel channel;
   // https://api.flutter.dev/javadoc/io/flutter/view/TextureRegistry.html
   private TextureRegistry texRegistry;
-  private Map<Long, SurfaceTextureEntry> textures;
+  private Map<Long, TextureEntry> textures; // SurfaceProducer or SurfaceTextureEntry
   private Map<Long, Surface> surfaces;
+  private boolean impeller;
   @Override
   public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
+    Bundle metaData = null;
+    Context appCtx = flutterPluginBinding.getApplicationContext();
+    try {
+        metaData = appCtx.getPackageManager().getApplicationInfo(appCtx.getPackageName(), PackageManager.GET_META_DATA).metaData;
+    } catch (PackageManager.NameNotFoundException e) {
+        throw new RuntimeException(e);
+    }
+    impeller = false;
+    if (metaData != null) {
+      impeller = metaData.getBoolean("io.flutter.embedding.android.EnableImpeller", false);
+    }
+    Log.i("FvpPlugin", "metaData: " + metaData + ", impeller: " + impeller);
     channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "fvp");
     channel.setMethodCallHandler(this);
     texRegistry = flutterPluginBinding.getTextureRegistry();
@@ -50,30 +68,69 @@ public class FvpPlugin implements FlutterPlugin, MethodCallHandler {
       final int width = (int)call.argument("width");
       final int height = (int)call.argument("height");
       final boolean tunnel = (boolean)call.argument("tunnel");
-      SurfaceTextureEntry te = texRegistry.createSurfaceTexture();
-      SurfaceTexture tex = te.surfaceTexture();
-      tex.setDefaultBufferSize(width, height); // TODO: size from player. rotate, fullscreen change?
-      Surface surface = new Surface(tex); // TODO: when to release
-      long texId = te.id();
+      TextureEntry te = null;
+      Surface surface = null;
+      final SurfaceProducer sp = impeller ? texRegistry.createSurfaceProducer() : null;
+      if (sp != null) {
+        sp.setSize(width, height);
+        surface = sp.getSurface();
+        te = sp;
+      } else {
+        SurfaceTextureEntry ste = texRegistry.createSurfaceTexture();
+        SurfaceTexture tex = ste.surfaceTexture();
+        tex.setDefaultBufferSize(width, height);
+        surface = new Surface(tex);
+        te = ste;
+      }
+      final long texId = te.id();
       nativeSetSurface(handle, texId, surface, width, height, tunnel);
       textures.put(texId, te);
       surfaces.put(texId, surface);
       result.success(texId);
+//// FLUTTER_3.24_BEGIN
+      if (sp != null) { // FIXME: requires 3.24. how to build conditionally?
+        // 3.24: https://docs.flutter.dev/release/breaking-changes/android-surface-plugins
+        sp.setCallback(
+                new TextureRegistry.SurfaceProducer.Callback() {
+                  @Override
+                  public void onSurfaceAvailable() {
+                    Log.d("FvpPlugin", "SurfaceProducer.onSurfaceAvailable for textureId " + texId);
+                    final Surface newSurface = sp.getSurface();
+                    surfaces.put(texId, newSurface);
+                    // will do nothing if same surface
+                    nativeSetSurface(handle, texId, newSurface, width, height, tunnel);
+                  }
+
+                  @Override
+                  public void onSurfaceCleanup() {
+                    Log.d("FvpPlugin", "SurfaceProducer.onSurfaceCleanup for textureId " + texId);
+                    textures.remove(texId);
+                    nativeSetSurface(handle, texId, null, 0, 0, tunnel);
+                  }
+                }
+        );
+      }
+//// FLUTTER_3.24_END
     } else if (call.method.equals("ReleaseRT")) {
       final int texId = call.argument("texture"); // 32bit int, 0, 1, 2 .... but SurfaceTexture.id() is long
       final long texId64 = texId; // MUST cast texId to long, otherwise remove() error
       nativeSetSurface(0, texId, null, -1, -1, false);
-      SurfaceTextureEntry te = textures.get(texId64);
+      TextureEntry te = textures.get(texId64);
       if (te == null) {
         Log.w("FvpPlugin", "onMethodCall: ReleaseRT texId not found: " + texId);
       } else {
         te.release();
       }
       if (textures.remove(texId64) == null) {
+        Log.w("FvpPlugin", "onMethodCall: ReleaseRT texture not found for " + texId);
       }
       if (surfaces.remove(texId64) == null) {
+        Log.w("FvpPlugin", "onMethodCall: ReleaseRT surface not found for " + texId);
       }
-      Log.w("FvpPlugin", "onMethodCall: ReleaseRT texId: " + texId + ", surfaces: " + surfaces.size() + " textures: " + textures.size());
+      Log.i("FvpPlugin", "onMethodCall: ReleaseRT texId: " + texId + ", surfaces: " + surfaces.size() + " textures: " + textures.size());
+      result.success(null);
+    } else if (call.method.equals("MixWithOthers")) {
+      // TODO: Implement actual business.
       result.success(null);
     } else {
       result.notImplemented();