+++ darwin/Classes/FvpPlugin.mm
@@ -1,12 +1,16 @@
-// Copyright 2023 Wang Bin. All rights reserved.
+// Copyright 2023-2025 Wang Bin. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#define USE_TEXCACHE 0
+
 #import "FvpPlugin.h"
-#import "Metal/Metal.h"
-#import "CoreVideo/CoreVideo.h"
 #include "mdk/RenderAPI.h"
 #include "mdk/Player.h"
+#import <AVFoundation/AVFoundation.h>
+#import <CoreVideo/CoreVideo.h>
+#import <Metal/Metal.h>
+#include <mutex>
 #include <unordered_map>
 #include <iostream>
 
@@ -22,7 +26,9 @@ using namespace std;
     id<MTLCommandQueue> cmdQueue;
     id<MTLTexture> texture;
     CVPixelBufferRef pixbuf;
+    id<MTLTexture> fltex;
     CVMetalTextureCacheRef texCache;
+    mutex mtx; // ensure whole frame render pass commands are recorded before blitting
 }
 
 - (instancetype)initWithWidth:(int)width height:(int)height
@@ -30,24 +36,48 @@ using namespace std;
     self = [super init];
     device = MTLCreateSystemDefaultDevice();
     cmdQueue = [device newCommandQueue];
-    CVMetalTextureCacheCreate(nullptr, nullptr, device, nullptr, &texCache);
+    auto td = [MTLTextureDescriptor texture2DDescriptorWithPixelFormat:MTLPixelFormatBGRA8Unorm width:width height:height mipmapped:NO];
+    td.usage = MTLTextureUsageRenderTarget;
+    texture = [device newTextureWithDescriptor:td];
+    //assert(!texture.iosurface); // CVPixelBufferCreateWithIOSurface(fltex.iosurface)
     auto attr = CFDictionaryCreateMutable(kCFAllocatorDefault, 0, &kCFTypeDictionaryKeyCallBacks, &kCFTypeDictionaryValueCallBacks);
     CFDictionarySetValue(attr, kCVPixelBufferMetalCompatibilityKey, kCFBooleanTrue);
+    auto iosurface_props = CFDictionaryCreateMutable(kCFAllocatorDefault, 0, &kCFTypeDictionaryKeyCallBacks, &kCFTypeDictionaryValueCallBacks);
+    CFDictionarySetValue(attr, kCVPixelBufferIOSurfacePropertiesKey, iosurface_props); // optional?
     CVPixelBufferCreate(nil, width, height, kCVPixelFormatType_32BGRA, attr, &pixbuf);
     CFRelease(attr);
+    texCache = {};
+#if (USE_TEXCACHE + 0)
+    CVMetalTextureCacheCreate(nullptr, nullptr, device, nullptr, &texCache);
     CVMetalTextureRef cvtex;
     CVMetalTextureCacheCreateTextureFromImage(nil, texCache, pixbuf, nil, MTLPixelFormatBGRA8Unorm, width, height, 0, &cvtex);
-    texture = CVMetalTextureGetTexture(cvtex);
+    fltex = CVMetalTextureGetTexture(cvtex);
     CFRelease(cvtex);
+#else
+    auto iosurface = CVPixelBufferGetIOSurface(pixbuf);
+    td.usage = MTLTextureUsageShaderRead; // Unknown?
+// macos: failed assertion `Texture Descriptor Validation IOSurface textures must use MTLStorageModeManaged or MTLStorageModeShared'
+// ios: failed assertion `Texture Descriptor Validation IOSurface textures must use MTLStorageModeShared
+    fltex = [device newTextureWithDescriptor:td iosurface:iosurface plane:0];
+#endif
     return self;
 }
 
 - (void)dealloc {
     CVPixelBufferRelease(pixbuf);
-    CFRelease(texCache);
+    if (texCache)
+        CFRelease(texCache);
 }
 
 - (CVPixelBufferRef _Nullable)copyPixelBuffer {
+    //return CVPixelBufferRetain(pixbuf);
+    scoped_lock lock(mtx);
+    auto cmdbuf = [cmdQueue commandBuffer];
+    auto blit = [cmdbuf blitCommandEncoder];
+    [blit copyFromTexture:texture sourceSlice:0 sourceLevel:0 sourceOrigin:MTLOriginMake(0, 0, 0) sourceSize:MTLSizeMake(texture.width, texture.height, texture.depth)
+        toTexture:fltex destinationSlice:0 destinationLevel:0 destinationOrigin:MTLOriginMake(0, 0, 0)]; // macos 10.15
+    [blit endEncoding];
+	[cmdbuf commit];
     return CVPixelBufferRetain(pixbuf);
 }
 @end
@@ -64,11 +94,13 @@ public:
         MetalRenderAPI ra{};
         ra.device = (__bridge void*)mtex_->device;
         ra.cmdQueue = (__bridge void*)mtex_->cmdQueue;
+// TODO: texture pool to avoid blitting
         ra.texture = (__bridge void*)mtex_->texture;
         setRenderAPI(&ra);
         setVideoSurfaceSize(width, height);
 
         setRenderCallback([this, texReg](void* opaque){
+            scoped_lock lock(mtex_->mtx);
             renderVideo();
             [texReg textureFrameAvailable:texId_];
         });
@@ -98,9 +130,16 @@ private:
     auto messenger = registrar.messenger;
 #else
     auto messenger = [registrar messenger];
+  // Allow audio playback when the Ring/Silent switch is set to silent
+    [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback error:nil];
 #endif
     FlutterMethodChannel* channel = [FlutterMethodChannel methodChannelWithName:@"fvp" binaryMessenger:messenger];
     FvpPlugin* instance = [[FvpPlugin alloc] initWithRegistrar:registrar];
+#if TARGET_OS_OSX
+#else
+  [registrar addApplicationDelegate:instance];
+#endif
+    [registrar publish:instance];
     [registrar addMethodCallDelegate:instance channel:channel];
     SetGlobalOption("MDK_KEY", "C03BFF5306AB39058A767105F82697F42A00FE970FB0E641D306DEFF3F220547E5E5377A3C504DC30D547890E71059BC023A4DD91A95474D1F33CA4C26C81B0FC73B00ACF954C6FA75898EFA07D9680B6A00FDF179C0A15381101D01124498AF55B069BD4B0156D5CF5A56DEDE782E5F3930AD47C8F40BFBA379231142E31B0F");
 }
@@ -118,8 +157,8 @@ private:
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
     if ([call.method isEqualToString:@"CreateRT"]) {
         const auto handle = ((NSNumber*)call.arguments[@"player"]).longLongValue;
-        const auto width = (int)((NSNumber*)call.arguments[@"width"]).longLongValue;
-        const auto height = (int)((NSNumber*)call.arguments[@"height"]).longLongValue;
+        const auto width = ((NSNumber*)call.arguments[@"width"]).intValue;
+        const auto height = ((NSNumber*)call.arguments[@"height"]).intValue;
         auto player = make_shared<TexturePlayer>(handle, width, height, _texRegistry);
         players[player->textureId()] = player;
         result(@(player->textureId()));
@@ -128,9 +167,31 @@ private:
         [_texRegistry unregisterTexture:texId];
         players.erase(texId);
         result(nil);
+    } else if ([call.method isEqualToString:@"MixWithOthers"]) {
+        [[maybe_unused]] const auto value = ((NSNumber*)call.arguments[@"value"]).boolValue;
+#if TARGET_OS_OSX
+#else
+        if (value) {
+            [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback withOptions:AVAudioSessionCategoryOptionMixWithOthers error:nil];
+        } else {
+            [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback error:nil];
+        }
+#endif
+        result(nil);
     } else {
         result(FlutterMethodNotImplemented);
     }
 }
 
+// ios only, optional. called first in dealloc(texture registry is still alive). plugin instance must be registered via publish
+- (void)detachFromEngineForRegistrar:(NSObject<FlutterPluginRegistrar> *)registrar {
+  players.clear();
+}
+
+#if TARGET_OS_OSX
+#else
+- (void)applicationWillTerminate:(UIApplication *)application {
+  players.clear();
+}
+#endif
 @end